Update Buy and Sell Design
